requirements.txt
boto3==1.40.11
fastapi==0.116.1
jinja2==3.1.6
pillow>=11.3.0
python-dotenv==1.1.1
python-multipart==0.0.20
uvicorn==0.35.0
